Multiple machines in same directory
You can now have multiple machines in the same directory.
Their HTML frags file must be named MACHINE-frags.html, and if
they have head-frags, those must be named MACHINE-head-frags.html.
The old syntax frags.html and head-frags.html is no longer recognized.

We also fixed a bug with the websocket upgrade so that it works
even if the port number is not the default.